# Optimize histogram to record single values