--- a/aggregators/codec_test.go
+++ b/aggregators/codec_test.go
@@ -65,7 +65,12 @@ func TestHistogramRepresentation(t *testing.T) {
 
 	actual := hdrhistogram.New()
 	HistogramFromProto(actual, HistogramToProto(expected))
-	assert.Empty(t, cmp.Diff(expected, actual))
+	assert.Empty(t, cmp.Diff(
+		expected, actual,
+		cmp.Comparer(func(a, b hdrhistogram.HybridCountsRep) bool {
+			return a.Equal(&b)
+		}),
+	))
 }
 
 func BenchmarkCombinedMetricsEncoding(b *testing.B) {