Instead of allocating values in a map for the first APMEvent received, we record the value as integers in the hybrid representation. The complexity of the approach is wrapped in a set of methods for ease of use.

Benchmarks:

```
name                        old time/op    new time/op    delta
CombinedMetricsEncoding-10    5.73µs ± 1%    5.70µs ± 2%     ~     (p=0.090 n=9+10)
CombinedMetricsDecoding-10    6.38µs ± 3%    5.12µs ± 4%  -19.84%  (p=0.000 n=10+10)
CombinedMetricsToBatch-10      352µs ± 1%     364µs ± 1%   +3.20%  (p=0.000 n=10+9)
EventToCombinedMetrics-10     1.25µs ± 2%    1.13µs ± 3%  -10.09%  (p=0.000 n=10+9)

name                        old alloc/op   new alloc/op   delta
CombinedMetricsEncoding-10    2.77kB ± 0%    2.77kB ± 0%     ~     (all equal)
CombinedMetricsDecoding-10    13.1kB ± 0%    10.3kB ± 0%  -21.93%  (p=0.000 n=8+9)
CombinedMetricsToBatch-10     39.7kB ± 0%    39.7kB ± 0%     ~     (all equal)
EventToCombinedMetrics-10     3.42kB ± 0%    3.14kB ± 0%   -8.44%  (p=0.000 n=10+10)

name                        old allocs/op  new allocs/op  delta
CombinedMetricsEncoding-10      79.0 ± 0%      79.0 ± 0%     ~     (all equal)
CombinedMetricsDecoding-10      80.0 ± 0%      40.0 ± 0%  -50.00%  (p=0.000 n=10+10)
CombinedMetricsToBatch-10        339 ± 0%       339 ± 0%     ~     (all equal)
EventToCombinedMetrics-10       19.0 ± 0%      15.0 ± 0%  -21.05%  (p=0.000 n=10+10)
```